requirements.txt
torch
transformers
pillow
opencv-python
ftfy
regex
tqdm
git+https://github.com/openai/CLIP.git
